refactor: TableGroup과 OrderTable을 다른 에그리거트로 분리
