+++ src/main/java/kitchenpos/ordertable/repository/TableGroupRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.ordertable.repository;
 
-import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
