+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -1,10 +1,10 @@
-package kitchenpos.ordertable.application;
+package kitchenpos.tablegroup.application;
 
 import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.ordertable.application.dto.TableGroupCreateRequest;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTables;
-import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.ordertable.repository.OrderTableRepository;
 import kitchenpos.ordertable.repository.TableGroupRepository;
 import org.springframework.stereotype.Service;
