+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -4,8 +4,8 @@ import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.ordertable.repository.OrderTableRepository;
 import kitchenpos.ordertable.repository.TableGroupRepository;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.TableGroup;
-import kitchenpos.ordertable.application.TableGroupService;
+import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.application.TableGroupService;
 import kitchenpos.ordertable.application.dto.OrderTableId;
 import kitchenpos.ordertable.application.dto.TableGroupCreateRequest;
 import org.junit.jupiter.api.DisplayName;
